mainRequests, searchRequests 하나로 합침
